@@ -1,13 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ApplicationIcon>mb256.ico</ApplicationIcon>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="CocApi.Cache" Version="2.0.0-preview1.13.4" />
+    <PackageReference Include="CocApi.Cache" Version="2.0.0-preview1.17.6" />
     <PackageReference Include="Serilog.Extensions.Hosting" Version="4.1.2" />
     <PackageReference Include="Serilog.Settings.Configuration" Version="3.3.0-dev-00291" />
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.1-dev-00947" />
